# src/main/resources/api.yaml
openapi: "3.0.3"
info:
  title: Learnify API
  description: API for Learnify application
  version: "1.0.0"
servers:
  - url: /api/v1
    description: Local server
paths:
  /workspaces:
    get:
      summary: List all workspaces owned by the user making the request
      operationId: listWorkspaces
      tags:
        - workspaces
      responses:
        '200':
          description: A list of workspaces
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/WorkspaceSummaryDto'
    post:
      summary: Create a workspace
      operationId: createWorkspace
      tags:
        - workspaces
      requestBody:
        description: Details to create a workspace
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WorkspaceCreateDto'
      responses:
        '201':
          description: Workspace created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkspaceSummaryDto'
  /notes:
    post:
      summary: Create a note
      operationId: createNote
      tags:
        - notes
      requestBody:
        description: Note details to create
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NoteCreateDto'
      responses:
        '201':
          description: Created note
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NoteSummaryDto'
  /notes/recent: #todo we should remeber to add userId when it will exists
    get:
      summary: List recent notes (max 4)
      operationId: listRecentNotes
      tags:
        - notes
      responses:
        '200':
          description: A list of recent notes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/NoteSummaryDto'
        '403':
          description: User has no permission to the other user's note
  /notes/{noteId}:
    get:
      summary: Info for a specific note
      operationId: getNoteById
      tags:
        - notes
      parameters:
        - $ref: '#/components/parameters/noteId'
      responses:
        '200':
          description: A single note details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NoteSummaryDto'
        '400':
          description: Bad request
        '403':
          description: User do not have permissions to get note details
        '404':
          description: Note not found
  /notes/{noteId}/board/pages/{pageNumber}:
    get:
      summary: Content from a specific page of board note
      operationId: getBoardNotePage
      tags:
        - notes
      parameters:
        - $ref: '#/components/parameters/noteId'
        - $ref: '#/components/parameters/pageNumber'
      responses:
        '200':
          description: Content of a single page board note
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BoardNotePageDto'
        '400':
          description: Bad request
        '403':
          description: User do not have permissions to get board page
        '404':
          description: Note or page not found
    put:
      summary: Update content of a specific page of board note
      operationId: updateBoardNotePage
      tags:
        - notes
      parameters:
        - $ref: '#/components/parameters/noteId'
        - $ref: '#/components/parameters/pageNumber'
      requestBody:
        description: Content to update
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BoardNotePageDto'
      responses:
        '200':
          description: Content updated
          content:
            application/json:
              schema:
                type: string
        '400':
          description: Bad request
        '403':
          description: User do not have permissions to update board page
  /notes/{noteId}/document/pages/{pageNumber}:
    get:
      summary: Content from a specific page of document note
      operationId: getDocumentNotePage
      tags:
        - notes
      parameters:
        - $ref: '#/components/parameters/noteId'
        - $ref: '#/components/parameters/pageNumber'
      responses:
        '200':
          description: Content of a single page document note
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DocumentNotePageDto'
        '400':
          description: Bad request
        '403':
          description: User do not have permissions to get document page
        '404':
          description: Note or page not found
    put:
      summary: Update content of a specific page of document note
      operationId: updateDocumentNotePage
      tags:
        - notes
      parameters:
        - $ref: '#/components/parameters/noteId'
        - $ref: '#/components/parameters/pageNumber'
      requestBody:
        description: Content to update
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DocumentNotePageDto'
      responses:
        '200':
          description: Content updated
          content:
            application/json:
              schema:
                type: string
        '400':
          description: Bad request
        '403':
          description: User do not have permissions to update document page
  /quizzes:
    post:
      summary: Create a quiz
      operationId: createQuiz
      tags:
        - quizzes
      requestBody:
        description: Quiz to create
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/QuizCreationDto'
      responses:
        '201':
          description: Quiz created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QuizDetailsDto'
  /quizzes/recent: #todo same as above, it should be done per user
    get:
      summary: List recent quizzes (max 4)
      operationId: listRecentQuizzes
      tags:
        - quizzes
      responses:
        '200':
          description: A list of recent quizzes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/QuizSummaryDto'
  /quizzes/{quizId}:
    get:
      summary: Info for a specific quiz
      operationId: showQuizById
      tags:
        - quizzes
      parameters:
        - name: quizId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: A single quiz
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QuizSummaryDto'
        '400':
          description: Bad request
        '403':
          description: User do not have permission to get resource
        '404':
          description: Quiz not found
  /quizzes/{quizId}/details:
    get:
      summary: Details of a specific quiz
      operationId: showDetailsQuizById
      tags:
        - quizzes
      parameters:
        - name: quizId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Details of a single Quiz
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QuizDetailsDto'
        '400':
          description: Bad request
        '403':
          description: User do not have permission to get resource
        '404':
          description: Not found details
  /quizzes/{quizId}/results:
    put:
      summary: Update results of a specific quiz
      operationId: updateResultsByQuizId
      tags:
        - quizzes
      parameters:
        - name: quizId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        description: Number of correct and incorrect answers of a quiz
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/QuizResultUpdateDto'
      responses:
        '200':
          description: Update results of a single Quiz
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QuizResultUpdateDto'
        '400':
          description: Bad request
        '403':
          description: User do not have permissions to get quiz results
        '404':
          description: Quiz not found
  /quizzes/{quizId}/questions:
    post:
      summary: Create questions
      operationId: createQuestions
      tags:
        - quizzes
      parameters:
        - name: quizId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        description: Questions to create
        required: true
        content:
          application/json:
            schema:
              type: array
              items:
                $ref: '#/components/schemas/QuestionCreationDto'
      responses:
        '201':
          description: Questions created
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/QuestionDto'
    get:
      summary: Questions for quiz
      operationId: showQuestionsByQuizId
      tags:
        - quizzes
      parameters:
        - name: quizId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Questions of a single Quiz
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/QuestionDto'
        '400':
          description: Bad request
        '403':
          description: User do not have permissions to get quiz questions
  /quizzes/{quizId}/questions/{questionId}:
    put:
      summary: Update question
      operationId: updateQuestion
      tags:
        - quizzes
      parameters:
        - name: quizId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: questionId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        description: Question to update
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/QuestionDto'
      responses:
        '200':
          description: Question updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QuestionDto'
        '400':
          description: Bad request
        '403':
          description: User do not have permissions to update quiz question
        '404':
          description: Quiz not found
  /bindings:
    post:
      summary: Create a binding between existing quiz and note
      operationId: createBinding
      tags:
        - bindings
      requestBody:
        description: Binding to create
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BindingCreateDto'
      responses:
        '201':
          description: Binding created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BindingDto'
  /bindings/notes/{noteId}:
    get:
      summary: List all quizzes bound to a note
      operationId: listQuizzesBoundToNote
      tags:
        - bindings
      parameters:
        - name: noteId
          in: path
          required: true
          schema:
            type: string
            format: uuid
            example: "sample00-id-0000-0000-note00000000"
            description: "UUID of the note"
      responses:
        '200':
          description: A list of quizzes bound to a note
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/QuizSummaryDto'
        '400':
          description: Bad request
        '403':
          description: User do not have permissions to get note binding
        '404':
          description: Note not found
  /bindings/quizzes/{quizId}:
    get:
      summary: List all notes bound to a quiz
      operationId: listNotesBoundToQuiz
      tags:
        - bindings
      parameters:
        - name: quizId
          in: path
          required: true
          schema:
            type: string
            format: uuid
            example: "sample00-id-0000-0000-quiz00000000"
            description: "UUID of the quiz"
      responses:
        '200':
          description: A list of notes bound to a quiz
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/NoteSummaryDto'
        '400':
          description: Bad request
        '403':
          description: User do not have permissions to get quiz binding
        '404':
          description: Quiz not found
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    WorkspaceCreateDto:
      type: object
      description: "Details to create a workspace"
      properties:
        displayName:
          type: string
          description: "Display name of the workspace"
          example: "Semestr 6"
      required: [ displayName ]
    NoteCreateDto:
      type: object
      description: "Details to create a note"
      properties:
        title:
          type: string
          description: "Title of the note"
          example: "Systemy rozproszone - middleware"
        description:
          type: string
          description: "Description of the note"
          example: "Ice/grpc/Thrift"
        workspaceId:
          type: string
          format: uuid
          description: "UUID of workspace"
          example: "sample00-id-0000-0000-workspaceid0"
        type:
          allOf:
            - $ref: '#/components/schemas/NoteTypeDto'
          example: "board"
      required: [ title, description, workspaceId, type ]
    NoteSummaryDto:
      type: object
      description: "Summary of a note - without content"
      properties:
        id:
          description: "UUID of the note"
          example: "sample00-id-0000-0000-note00000000"
          type: string
          format: uuid
        workspace:
          $ref: '#/components/schemas/WorkspaceSummaryDto'
        title:
          type: string
          description: "Title of the note"
          example: "Systemy rozproszone - middleware"
        description:
          type: string
          description: "Description of the note"
          example: "Ice/grpc/Thrift"
        author:
          $ref: '#/components/schemas/UserSummaryDto'
        updatedAt:
          description: "Date of last modification"
          type: string
          format: date-time
        createdAt:
          description: "Date of creation"
          type: string
          format: date-time
        type:
          allOf:
            - $ref: '#/components/schemas/NoteTypeDto'
          example: "board"
      required: [ id, workspace, title, description, author, updatedAt, createdAt, type ]


    # TODO decide if we should store more detailed objects on backend side or just store the content as string (json)
    # storing as string is easier to implement but harder to use on backend - such like allowing PATCH operation on content
    # on the other hand storing as object introduces many mappings between layers what is not the best idea with saving data very often

    BoardNotePageDto:
      type: object
      description: "Content of a single of board page note"
      properties:
        content:
          type: string
          description: "Content of the note"
          example: "This is a sample content of the note"
      required: [ content ]

    DocumentNotePageDto:
      type: object
      description: "Content of a single of text document page note"
      properties:
        content:
          type: string
          description: "Content of the note"
          example: "This is a sample content of the note"
      required: [ content ]

    WorkspaceSummaryDto:
      type: object
      description: "Summary of a workspace"
      required:
        - id
        - displayName
      properties:
        id:
          type: string
          format: uuid
          description: "UUID of the workspace"
          example: "sample00-id-0000-0000-workspace000"
        displayName:
          type: string
          description: "Display name of the workspace"
          example: "Semestr 6"
    UserSummaryDto:
      type: object
      description: "Summary of a user"
      required:
        - id
        - displayName
        - email
      properties:
        id:
          type: string
          description: "ID of the user"
          example: "asdoifihgausyiodqfeuvhk123"
        displayName:
          type: string
          description: "Display name of the user"
          example: "Krzysztof Usnarski"
        email:
          type: string
          description: "User's email"
          example: "testUser@gmail.com"
    QuizSummaryDto:
      type: object
      required:
        - id
        - workspace
        - author
        - title
        - score
      properties:
        id:
          type: string
          format: uuid
        workspace:
          $ref: '#/components/schemas/WorkspaceSummaryDto'
        title:
          type: string
        score:
          type: string
        author:
          $ref: '#/components/schemas/UserSummaryDto'
        date:
          type: string

    QuizSimpleResultDto:
      type: object
      required:
        - incorrect
        - correct
      properties:
        incorrect:
          type: integer
          format: int32
        correct:
          type: integer
          format: int32
    QuizCreationDto:
      type: object
      required:
        - workspaceId
        - title
        - description
      properties:
        workspaceId:
          type: string
          format: uuid
        title:
          type: string
        description:
          type: string
    QuizDetailsDto:
      type: object
      required:
        - id
        - workspace
        - title
        - description
        - numberOfQuestions
        - author
        - createdAt
      properties:
        id:
          type: string
          format: uuid
        workspace:
          $ref: '#/components/schemas/WorkspaceSummaryDto'
        title:
          type: string
        description:
          type: string
        numberOfQuestions:
          type: integer
          format: int32
        lastScore:
          $ref: '#/components/schemas/QuizSimpleResultDto'
        bestScore:
          $ref: '#/components/schemas/QuizSimpleResultDto'
        author:
          $ref: '#/components/schemas/UserSummaryDto'
        lastTryDate:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
    QuestionDto:
      type: object
      required:
        - questionId
        - question
        - type
        - quizId
        - weight
        - choices
        - feedback
        - otherProperties
      properties:
        questionId:
          type: string
          format: uuid
        question:
          type: string
        type:
          allOf:
            - $ref: '#/components/schemas/QuestionTypeDto'
        quizId:
          type: string
          format: uuid
        weight:
          type: integer
          format: int32
        choices:
          type: string
        feedback:
          type: string
        otherProperties:
          type: string
    QuestionCreationDto:
      type: object
      required:
        - question
        - type
        - weight
        - choices
        - feedback
        - otherProperties
      properties:
        question:
          type: string
        type:
          allOf:
            - $ref: '#/components/schemas/QuestionTypeDto'
        weight:
          type: integer
          format: int32
        choices:
          type: string
        feedback:
          type: string
        otherProperties:
          type: string
    QuizResultUpdateDto:
      type: object
      required:
        - correct
        - incorrect
      properties:
        correct:
          type: integer
          format: int32
        incorrect:
          type: integer
          format: int32
    BindingCreateDto:
      type: object
      required:
        - quizId
        - noteId
      properties:
        quizId:
          type: string
          format: uuid
        noteId:
          type: string
          format: uuid
    BindingDto:
      type: object
      required:
        - bindingId
        - quizId
        - noteId
      properties:
        bindingId:
          type: string
          format: uuid
        quizId:
          type: string
          format: uuid
        noteId:
          type: string
          format: uuid
    NoteTypeDto:
      type: string
      enum: [ document, board ]
      example: document
    QuestionTypeDto:
      type: string
      enum: [ multiple-choice, single-choice ]
      example: multiple-choice

  parameters:
    noteId:
      name: noteId
      in: path
      required: true
      description: "UUID of the note"
      schema:
        type: string
        format: uuid
        example: "sample00-id-0000-0000-note00000000"
    pageNumber:
      name: pageNumber
      in: path
      required: true
      description: "Page number"
      schema:
        type: integer
        format: int32
        example: 1

security:
    - bearerAuth: []